+++ libspaghetti/include/spaghetti/node.h
@@ -107,9 +107,9 @@ class SPAGHETTI_API Node : public QGraphicsItem {
 
   void changeInputName(int const a_id, QString const &a_name);
   void changeOutputName(int const a_id, QString const &a_name);
-  void addSocket(SocketType const a_type, uint8_t const a_id, QString const &a_name, ValueType const a_valueType);
+  void addSocket(SocketType const a_type, uint64_t const a_id, QString const &a_name, ValueType const a_valueType);
   void removeSocket(SocketType const a_type);
-  void setSocketType(IOSocketsType const a_socketType, uint8_t const a_socketId, ValueType const a_type);
+  void setSocketType(IOSocketsType const a_socketType, uint64_t const a_socketId, ValueType const a_type);
 
  protected:
   void setCentralWidget(QGraphicsItem *a_centralWidget);
@@ -119,11 +119,11 @@ class SPAGHETTI_API Node : public QGraphicsItem {
  private:
   void addInput();
   void removeInput();
-  void setInputName(uint8_t const a_socketId, QString const &a_name);
+  void setInputName(uint64_t const a_socketId, QString const &a_name);
 
   void addOutput();
   void removeOutput();
-  void setOutputName(uint8_t const a_socketId, QString const &a_name);
+  void setOutputName(uint64_t const a_socketId, QString const &a_name);
 
   void updateOutputs();
